\name{data_ace}
\alias{data_ace}
\docType{data}
\title{
Example data for the ACEt models
}
\description{
This is an example dataset consisting of traits and ages for MZ and DZ twins.
}
\usage{data(data_ace)}
\format{
  The format is:

}

\examples{
data(data_ace)
}
\keyword{datasets}
