\name{data_ace}
\alias{data_ace}
\docType{data}
\title{
Example data for the ACEt models
}
\description{
This is an example dataset consisting of traits and ages for MZ and DZ twins.
}
\usage{data(data_ace)}
\format{
  The format is:
	mz: a matrix of simulated data for MZ twins.
	dz: a matrix of simulated data for DZ twins.
}

\examples{
data(data_ace)
}
\keyword{datasets}
